# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "College Event Portal API",
    "version": "2.0.0",
    "description": "Endpoints for browsing events, registering, 2FA login, and admin/organizer management."
  },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Liveness check",
        "responses": { "200": { "description": "Server is up" } }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Create a student account",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["name", "email", "phone", "password"],
            "properties": {
              "name": { "type": "string" },
              "email": { "type": "string" },
              "phone": { "type": "string", "example": "+919876543210" },
              "password": { "type": "string" },
              "department": { "type": "string" }
            }
          } } }
        },
        "responses": { "200": { "description": "Account created, JWT returned" }, "400": { "description": "Validation error" }, "409": { "description": "Email or phone already exists" } }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Start login â€” sends a 2FA code by email",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["email", "password"],
          "properties": { "email": { "type": "string" }, "password": { "type": "string" } }
        } } } },
        "responses": { "200": { "description": "OTP sent" }, "401": { "description": "Invalid credentials" } }
      }
    },
    "/api/verify-otp": {
      "post": {
        "summary": "Complete login with the emailed 6-digit code",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["userId", "otpCode"],
          "properties": { "userId": { "type": "string" }, "otpCode": { "type": "string" } }
        } } } },
        "responses": { "200": { "description": "JWT returned" }, "400": { "description": "Invalid or expired code" } }
      }
    },
    "/api/forgot-password": {
      "post": {
        "summary": "Request a password reset code by email",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["email"], "properties": { "email": { "type": "string" } }
        } } } },
        "responses": { "200": { "description": "Reset code sent" }, "404": { "description": "No account with that email" } }
      }
    },
    "/api/reset-password": {
      "post": {
        "summary": "Set a new password using the emailed reset code",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["userId", "otpCode", "newPassword"],
          "properties": { "userId": { "type": "string" }, "otpCode": { "type": "string" }, "newPassword": { "type": "string" } }
        } } } },
        "responses": { "200": { "description": "Password reset" }, "400": { "description": "Invalid or expired code" } }
      }
    },
    "/api/events": {
      "get": {
        "summary": "List events (search/category/department filters, optional pagination)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "department", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "List of events" } }
      },
      "post": {
        "summary": "Create an event (organizer/admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Event created" }, "403": { "description": "Permission denied" } }
      }
    },
    "/api/events/{id}/register": {
      "post": {
        "summary": "Register for an event (student only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Registered, QR ticket returned" }, "400": { "description": "Event full" }, "409": { "description": "Already registered" } }
      }
    },
    "/api/events/{id}/waitlist": {
      "post": {
        "summary": "Join the waitlist for a full event (student only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Joined the waitlist" }, "400": { "description": "Event still has open seats" } }
      },
      "delete": {
        "summary": "Leave the waitlist for an event (student only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Removed from waitlist" }, "404": { "description": "Not on the waitlist" } }
      }
    },
    "/api/events/{id}/registrations/export": {
      "get": {
        "summary": "Export attendees for an event as CSV (organizer/admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "CSV file" }, "403": { "description": "Permission denied" } }
      }
    },
    "/api/scan": {
      "post": {
        "summary": "Verify a scanned QR ticket and mark attendance (organizer/admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Attendance marked" }, "400": { "description": "Invalid or expired ticket" } }
      }
    }
  }
}